+++ src/timber_design/ghpython/components/CT_Panel_Joint_Rule_Category/code.py
@@ -1,5 +1,4 @@
 # flake8: noqa
-import inspect
 from collections import OrderedDict
 
 import Grasshopper
@@ -13,6 +12,7 @@ from timber_design.ghpython import manage_cpython_dynamic_params
 from timber_design.ghpython import rename_cpython_gh_output
 from timber_design.ghpython import warning
 from timber_design.ghpython import message
+from timber_design.ghpython.joint_arg_mapping import get_gh_arg_names
 
 
 class CategoryPanelJointRule(Grasshopper.Kernel.GH_ScriptInstance):
@@ -44,10 +44,7 @@ class CategoryPanelJointRule(Grasshopper.Kernel.GH_ScriptInstance):
             return CategoryRule(self.joint_type, cat_a, cat_b)
 
     def arg_names(self):
-        names = inspect.getfullargspec(self.joint_type.__init__)[0][1:3]
-        for i in range(2):
-            names[i] += "_category"
-        return [name for name in names if (name != "key") and (name != "frame")] + ["max_distance"]
+        return get_gh_arg_names(self.joint_type, CategoryRule, expose_extra_kwargs=False)
 
     def AppendAdditionalMenuItems(self, menu):
         for name in self.classes.keys():