+++ src/timber_design/ghpython/components/CT_Plate_Joint_Rule_Direct/code.py
@@ -1,7 +1,5 @@
 """Generates a direct joint between two elements. This overrides other joint rules."""
 
-import inspect
-
 import compas_timber.connections as _ct_connections
 import Grasshopper  # type: ignore
 from compas_timber.connections import PanelJoint
@@ -12,6 +10,8 @@ from timber_design.ghpython import item_input_valid_cpython
 from timber_design.ghpython import manage_cpython_dynamic_params
 from timber_design.ghpython import rename_cpython_gh_output
 from timber_design.ghpython import warning
+from timber_design.ghpython.joint_arg_mapping import build_joint_kwargs
+from timber_design.ghpython.joint_arg_mapping import get_gh_arg_names
 from timber_design.workflow import DirectRule
 
 
@@ -54,11 +54,12 @@ class DirectPlateJointRule(Grasshopper.Kernel.GH_ScriptInstance):
         for i, val in enumerate(args[2:]):
             if val is not None:
                 kwargs[self.arg_names()[i + 2]] = val
+        kwargs = build_joint_kwargs(self.joint_type, kwargs, main_beam=plate_a, cross_beam=plate_b)
 
         return DirectRule(self.joint_type, [plate_a, plate_b], **kwargs)
 
     def arg_names(self):
-        return inspect.getargspec(self.joint_type.__init__)[0][1:3] + ["max_distance"]
+        return get_gh_arg_names(self.joint_type, DirectRule, expose_extra_kwargs=False)
 
     def AppendAdditionalMenuItems(self, menu):
         for name in self.classes.keys():